% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateExpPrior.R
\name{estimateExpPrior}
\alias{estimateExpPrior}
\title{Define/estimate normal multivariate prior pdf for exponential decay parameters.}
\usage{
estimateExpPrior(x, uy, dataType, priorType = "mono", out,
  ru_theta = 0.05, eps = 0.001, nb_chains = 4, nb_warmup = 800,
  nb_iter = nb_warmup + 200)
}
\arguments{
\item{x}{a numeric vector of depths}

\item{uy}{a numeric vector of uncertainties}

\item{dataType}{an numeric (1 or 2) defining the type of data}

\item{priorType}{a string defining the type of prior}

\item{out}{ourput from \code{fitMonoExp}}

\item{ru_theta}{optional real defining the relative uncertainty on parameters}

\item{eps}{tolerance parameter for moments matching method}

\item{nb_chains}{number of MCMC chains}

\item{nb_warmup}{number of warmup steps}

\item{nb_iter}{number of steps}
}
\value{
A list containing the center, covariance matrix 
of the prior pdf and the constraint and realized statistics.
}
\description{
Define/estimate normal multivariate prior pdf for exponential decay parameters.
}
\details{
Provides two ways to buil a prior for the exponential decay 
parameters of the \code{ExpGP} model:
\describe{
  \item{priorType='mono'}{builds a covariance matrix from the correlation 
    matrix of the \code{fitMonoExp} model and a relative uncertainty 
    parameter \code{ru_theta}}
  \item{priorType='abc'}{estimates the parameters uncertainties by 
    a moments matching strategy, and assume no correlation}
}
}
\author{
Pascal PERNOT
}
