% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNoise.R
\name{estimateNoise}
\alias{estimateNoise}
\title{Estimate and model noise in signal}
\usage{
estimateNoise(x, y, df = 15)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{df}{smoothing factor for \code{smooth.splines}}
}
\value{
A list containing
\describe{
  \item{fit}{a \code{stanfit} object containg the results of the fit}
  \item{method}{choice of optimization method}
  \item{theta}{a vector of optimal parameters}
  \item{uy}{a vector of estimated uncertainty values for \code{y}}
  \item{ySmooth}{ a vector of values for the smoother curve}
}
}
\description{
Estimate and model noise in signal
}
\details{
Function which proceeds in two steps:
\enumerate{
  \item get a set of residuals R using a smoothing splines model
  \item estimate the x-dependent standard deviation of the residuals
}
by bayesian inference: \code{R(x) ~ normal(0,uy(x));
uy(x) = theta[1]*exp(-x/theta[2])} assuming a Poisson-type noise.
}
\author{
Pascal PERNOT
}
