% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitExpGP.R
\name{fitExpGP}
\alias{fitExpGP}
\title{Decay fit with modulation of mean depth}
\usage{
fitExpGP(x, y, uy, Nn = 10, theta0 = NULL, cor_theta = NULL,
  ru_theta = 0.1, lambda_rate = 0.1, lasso = FALSE,
  method = "sample", iter = 50000, prior_PD = 0, nb_chains = 4,
  nb_warmup = 500, nb_iter = 1000)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{uy}{a numeric vector of uncertainty on y}

\item{method}{choice of optimization method in c('optim','sample')}
}
\value{
A list
}
\description{
Decay fit with modulation of mean depth
}
\details{
Function which proceeds in two steps:
}
\author{
Pascal PERNOT
}
