% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMonoExp.R
\name{fitMonoExp}
\alias{fitMonoExp}
\title{Monoexponential decay fit}
\usage{
fitMonoExp(x, y, uy, method = "optim", nb_chains = 4,
  nb_warmup = 500, nb_iter = nb_warmup + 500)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{uy}{a numeric vector of uncertainty on y}

\item{method}{choice of optimization method in c('optim','sample')}
}
\value{
A list
}
\description{
Monoexponential decay fit
}
\details{
Function which proceeds in two steps:

(1) get a set of residuals R using a smoothing splines model

(2) estimate the x-dependent standard deviation of the residuals

by bayesian inference: R(x) ~ normal(0,uy(x));

uy(x) = theta[1]*exp(-x/theta[2]) assuming a Poisson-type noise.
}
\author{
Pascal PERNOT
}
