% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitOCTLib-package.R
\docType{package}
\name{FitOCTLib-package}
\alias{FitOCTLib-package}
\alias{FitOCTLib}
\title{The 'FitOCTLib' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.1. http://mc-stan.org
}
