% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nActCtrlPts.R
\name{nActCtrlPts}
\alias{nActCtrlPts}
\title{Count active control points in fitExpGP output.}
\usage{
nActCtrlPts(fit, prob = 0.9)
}
\arguments{
\item{fit}{an object (list or stanfit) issued from a stan code}

\item{prob}{a probability threshold}
}
\value{
The number of active points or NULL.
}
\description{
Count active control points in fitExpGP output.
}
\details{
Parameter \code{prob} defines the probability
above which a control point is considered active by differing 
from zero.
}
\author{
Pascal PERNOT
}
