% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpGP.R
\name{plotExpGP}
\alias{plotExpGP}
\title{Plot outputs from \code{fitExpGP}.}
\usage{
plotExpGP(x, y, uy, ySmooth, out, modScale = 0.3, nMC = 100, gPars)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses/data}

\item{uy}{a numeric vector of uncertainties}

\item{ySmooth}{a numeric vector of smoothed data}

\item{out}{a list, output of fitEpxGP}

\item{modScale}{a real defining the plotting scale for yGP}

\item{nMC}{an integer number of spaghetti lines}

\item{gPars}{a list of graphical parameters and colors}
}
\value{
Produces a plot.
}
\description{
Plot outputs from \code{fitExpGP}.
}
\author{
Pascal PERNOT
}
