% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMonoExp.R
\name{plotMonoExp}
\alias{plotMonoExp}
\title{Plot outputs from \code{fitMonoExp}.}
\usage{
plotMonoExp(x, y, uy, ySmooth, mod, resid, gPars)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses/data}

\item{uy}{a numeric vector of uncertainties}

\item{ySmooth}{a numeric vector of smoothed data}

\item{mod}{a numeric vector of model values}

\item{resid}{a numeric vector of residuals}

\item{gPars}{a list of graphical parameters and colors}
}
\value{
Produces a plot.
}
\description{
Plot outputs from \code{fitMonoExp}.
}
\author{
Pascal PERNOT
}
