% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNoise.R
\name{plotNoise}
\alias{plotNoise}
\title{Plot outputs from \code{estimaNoise}.}
\usage{
plotNoise(x, y, uy, ySmooth, gPars)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{uy}{a numeric vector of uncertainties}

\item{ySmooth}{a numeric vector of smoothed data}

\item{gPars}{a list of graphical parameters and colors}
}
\value{
Produces a plot.
}
\description{
Plot outputs from \code{estimaNoise}.
}
\author{
Pascal PERNOT
}
