% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriPostAll.R
\name{plotPriPostAll}
\alias{plotPriPostAll}
\title{Plot matrix of marginal prior and posterior pdfs.}
\usage{
plotPriPostAll(pri, pos, gPars)
}
\arguments{
\item{pri}{a stanfit object with prior samples}

\item{pos}{a stanfit object with posterior samples}

\item{gPars}{a list of graphical parameters and colors}
}
\value{
Produces a plot.
}
\description{
Plot matrix of marginal prior and posterior pdfs.
}
\author{
Pascal PERNOT
}
