% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printBr.R
\name{printBr}
\alias{printBr}
\title{Print Birges's ratio and confidence interval.}
\usage{
printBr(fit, prob = 0.9, silent = FALSE)
}
\arguments{
\item{fit}{an object (list or stanfit) issued from a stan code}

\item{prob}{a probability threshold}

\item{silent}{a logical controling standard output}
}
\value{
Returns invisibly a list containing
\describe{
  \item{br}{the estimated value of the Birge's ratio} 
  \item{ndf}{the estimated number of degrees of freedom} 
  \item{CI95}{a 95\% interval for admissible values of br}
  \item{alert}{a possible warning message, or NULL}
}
}
\description{
Print Birges's ratio and confidence interval.
}
\details{
Birges's ratio and a 95% interavl is estimated for 
fitMonoExp or fitExpGP models. For the latter, the number of
degrees of freedom is based on a count of active control points
in the GP model. Parameter \code{prob} defines the probability
above which a control point is considered active by differing 
from zero.
}
\author{
Pascal PERNOT
}
