% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selX.R
\name{selX}
\alias{selX}
\title{Subset OCT signal.}
\usage{
selX(x, y, depthSel = NULL, subSample = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{depthSel}{range of \code{x} values: \code{c(xmin,xmax)}}

\item{subsample}{a numeric factor for regular subsampling}
}
\value{
Returns a list containing new vectors \code{x} and \code{y}.
}
\description{
Subset OCT signal.
}
\details{
The default parameters retuen the input vectors unchanged.
}
\author{
Pascal PERNOT
}
