% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateExpPrior.R
\name{estimateExpPrior}
\alias{estimateExpPrior}
\title{Define/estimate normal multivariate prior pdf for exponential decay parameters.}
\usage{
estimateExpPrior(
  x,
  uy,
  dataType = 2,
  priorType = "mono",
  out,
  ru_theta = 0.05,
  eps = 0.001,
  nb_chains = 4,
  nb_warmup = 800,
  nb_iter = nb_warmup + 200
)
}
\arguments{
\item{x}{numeric vector of depths}

\item{uy}{numeric vector of uncertainties}

\item{dataType}{integer defining the type of data (1:amplitude or 
2:intensity) )}

\item{priorType}{string defining the type of prior ('mono' or 'abc')}

\item{out}{output list from \code{fitMonoExp}}

\item{ru_theta}{optional positive real defining the relative uncertainty 
on parameters (priorType='mono')}

\item{eps}{tolerance parameter for the moments matching method 
(priorType='abc')}

\item{nb_chains}{number of MCMC chains (priorType='abc')}

\item{nb_warmup}{number of warmup steps (priorType='abc')}

\item{nb_iter}{number of steps (priorType='abc')}
}
\value{
A list containing: the center, covariance matrix 
        of the prior pdf and, for \code{priorType = 'abc'}, 
        a \code{stats} list containing the constraint \code{Sobs} 
        and realized \code{Ssim} statistics.
}
\description{
Define/estimate normal multivariate prior pdf for exponential decay parameters.
}
\details{
Provides two ways to buil a normal multivariate prior for 
         the exponential decay parameters of the \code{ExpGP} model.
         The mean value is in both cases the MAP issued by fitMonoExp.
         For the covariance matrix one has two options: 
         \describe{
            \item{priorType='mono'}{
            a covariance matrix is built from the correlation matrix 
            estimated by the \code{fitMonoExp} model and a relative 
            uncertainty on the parameters (\code{ru_theta}). 
            This accounts for the fact that the parameters 
            uncertainties provided by fitMonoExp are not reliable,
            as they are issued from an invalid model.
            }
            \item{priorType='abc'}{
            the parameters uncertainties/variances are optimized
            by a moments matching strategy: (1) the 2-sigma prediction 
            uncertainty has to match the 95-th quantile of the absolute
            errors of the fitMonoExp model (this statistics is weighted 
            by \code{uy}); and (2) the standard deviation of the prediction 
            uncertainty has to be as small as possible.
            This prior assumes a diagonal covariance matrix.
            }
         }
}
\author{
Pascal PERNOT
}
