% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateExpPrior.R
\name{estimateExpPrior}
\alias{estimateExpPrior}
\title{Define/estimate normal multivariate prior pdf for exponential decay parameters.}
\usage{
estimateExpPrior(x, uy, dataType, priorType = "mono", out,
  ru_theta = 0.05)
}
\arguments{
\item{x}{a numeric vector of depths}

\item{uy}{a numeric vector of uncertainties}

\item{dataType}{an numeric (1 or 2) defining the type of data}

\item{priorType}{a string defining the type of prior}

\item{out}{ourput from \code{fitMonoExp}}

\item{ru_theta}{optional real defining the relative uncertainty on parameters}
}
\value{
A list containing the center and covariance matrix of the prior pdf.
}
\description{
Define/estimate normal multivariate prior pdf for exponential decay parameters.
}
\author{
Pascal PERNOT
}
