% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNoise.R
\name{estimateNoise}
\alias{estimateNoise}
\title{Estimate and model noise in signal}
\usage{
estimateNoise(x, y, df = 15)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{df}{smoothing factor for `smooth.splines`}
}
\value{
A list
}
\description{
Estimate and model noise in signal
}
\details{
Function which proceeds in two steps:

(1) get a set of residuals R using a smoothing splines model

(2) estimate the x-dependent standard deviation of the residuals

by bayesian inference: R(x) ~ normal(0,uy(x));

uy(x) = theta[1]*exp(-x/theta[2]) assuming a Poisson-type noise.
}
\author{
Pascal PERNOT
}
