% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDecayModel.R
\name{expDecayModel}
\alias{expDecayModel}
\title{Reference OCT decay model.}
\usage{
expDecayModel(x, c)
}
\arguments{
\item{x}{a numeric vector of depths}

\item{c}{a numeric vector of parameters}
}
\value{
A numeric vector of model values
}
\description{
Reference OCT decay model.
}
\author{
Pascal PERNOT
}
