% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitExpGP.R
\name{fitExpGP}
\alias{fitExpGP}
\title{Decay fit with modulation of mean depth by Gaussian Process}
\usage{
fitExpGP(x, y, uy, dataType = 2, Nn = 10, theta0 = NULL,
  Sigma0 = NULL, lambda_rate = 0.1, lasso = FALSE,
  method = "sample", iter = 50000, prior_PD = 0, alpha_scale = 0.1,
  rho_scale = 1/Nn, gridType = "internal", nb_chains = 4,
  nb_warmup = 500, nb_iter = 1000, verbose = FALSE,
  open_progress = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{uy}{a numeric vector of uncertainty on 'y'}

\item{dataType}{an numeric (1 or 2) defining the type of data}

\item{Nn}{number of control points}

\item{theta0}{theta prior mean vbalues}

\item{Sigma0}{theta prior corvariance matrix}

\item{lambda_rate}{scale of ctrl points prior}

\item{lasso}{flag to use lasso prior}

\item{method}{choice of optimization method in c('optim','sample')}

\item{iter}{max. number of iterations for 'optim'}

\item{prior_PD}{flag to sample from prior pdf only}

\item{alpha_scale}{SD scale of GP}

\item{rho_scale}{relative correlation length of GP}

\item{gridType}{type of controle points grid (`internal` does not contain boundaries)}

\item{nb_chains}{number of MCMC chains}

\item{nb_warmup}{number of warmup steps}

\item{nb_iter}{number of steps}
}
\value{
A list containing
\describe{
  \item{fit}{a \code{stanfit} object containg the results of the fit}
  \item{xGP}{a vector of coordinates for the control points}
  \item{method}{same as input}
  \item{prior_PD}{same as input}
  \item{lasso}{same as input}
}
}
\description{
Decay fit with modulation of mean depth by Gaussian Process
}
\details{
Bayesian inference of the parameters of an exponential
model with modulation assuming an uncorrelated normal noise
\code{y(x) ~ normal(m(x),uy(x));
m(x) = theta[1] + theta[2]*exp(-2*x/theta[3]*(1+l(x)));}.
\code{l(x)} is defined by a GP with fixed positions 'xGP',
variance and correlation length.
}
\author{
Pascal PERNOT
}
