% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMonoExp.R
\name{fitMonoExp}
\alias{fitMonoExp}
\title{Monoexponential fit of OCT decay}
\usage{
fitMonoExp(x, y, uy, method = "optim", nb_chains = 4,
  nb_warmup = 500, nb_iter = nb_warmup + 500)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of responses}

\item{uy}{a numeric vector of uncertainty on 'y'}

\item{method}{choice of optimization method in c('optim','sample')}
}
\value{
A list containing
\describe{
  \item{best.theta}{a vector of optimal parameters}
  \item{cor.theta}{correlation matrix of optimal parameters}
  \item{method}{choice of optimization method}
  \item{fit}{a \code{stanfit} object containg the results of the fit}
}
}
\description{
Monoexponential fit of OCT decay
}
\details{
Bayesian inference of the parameters of an exponential
model assuming an uncorrelated normal noise
\code{y(x) ~ normal(m(x),uy(x));
m(x) = theta[1] + theta[2]*exp(-x/theta[3])}.
}
\author{
Pascal PERNOT
}
