% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriPost.R
\name{plotPriPost}
\alias{plotPriPost}
\title{Plot 1 marginal prior and posterior pdf.}
\usage{
plotPriPost(pri, pos, tag, xlim = range(c(pri, pos)), gPars, kl = TRUE)
}
\arguments{
\item{pri}{prior sample}

\item{pos}{posterior sample}

\item{tag}{name of the variable}

\item{xlim}{an interval}

\item{gPars}{a list of graphical parameters and colors}

\item{kl}{a logical to compute and display KL-divergence}
}
\value{
Produces a plot. The KL divergence in bits is
        displayed in the title if `kl == TRUE`.
}
\description{
Plot 1 marginal prior and posterior pdf.
}
\author{
Pascal PERNOT
}
